Refactor OTLP endpoint for extensibility to logs and traces
Extract shared OTLP logic into abstract base classes to prepare for
adding logs and traces support:

- AbstractOTLPTransportAction: bulk request lifecycle, success/partial-success/failure handling
- AbstractOTLPRestAction: protobuf content-type validation, request dispatching
- OTelDocumentBuilder: resource, scope, attributes, and data stream document building
- OTLPActionRequest/OTLPActionResponse: shared request/response types
- AbstractOTLPIndexingRestIT: shared test cluster setup
- Move extractReceiverName to TargetIndex
